Move tasks menu to new header
